[Build] Commit the DirectInput 7 import library
This is the only .lib file that is no longer shipped in the current
Windows SDK. If it was only for DirectInputCreate(), I would have just
gone the LoadLibrary()/GetProcAddress() route, but [c_dfDIKeyboard] is
a rather unwieldy piece of data to copy-paste.

Unfortunately, we can't also just quickly update this code to
DirectInput 8, as explained in #22.

Part of P0217, funded by Arandui.