[Graphics] Distinguish palettized (= 8-bit) and channeled (≥ 16-bit) bit depths
In total, the differences run way too deep to abstract them away nicely
within one push, but this is the core difference between most of the
custom code for either bit depth…

…except when the code actually intends to branch based on whether the
game renders in 2D ("8-bit") or hybrid 2D/3D ("16-bit") mode. But it's
hard to tell whether each piece of code branches based on actual bit
depth, the 2D/3D rendering approach, or both. This is fine for a start.

Part of P0226, funded by alp-bib and Arandui.