[Graphics] Support 32-bit rendering
Interestingly, DirectDraw indicates that it doesn't even support 8-bit
or 16-bit modes on my system. As it turns out, it only does if the
`DWM8And16BitMitigation` is enabled in either

	• `HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows NT\CurrentVersion\AppCompatFlags\Layers`, or
	• `HKEY_CURRENT_USER\SOFTWARE\Microsoft\Windows NT\CurrentVersion\AppCompatFlags\Layers`,

confirming the initial suspicion in #3. Said mitigation is exactly what
slows the game down to a crawl though… might want to fix that as well?

Closes #3.

Part of P0226, funded by alp-bib and Arandui.