[Config] Move file I/O to a new module
GIAN.CPP is *kind of* supposed to be this module, but GIAN.H is also a
giant mess.
Due to DEMOPLAY.H's dependency on CONFIG_DATA, we also need to remove
this inclusion from GIAN.H for this to actually cut down on the number
of recompiled translation units.

Part of P0226, funded by alp-bib and Arandui.