[Ending] Set the intended text font
As it turns out, systems with Japanese locale automatically select MS
Gothic at 16 pixels as the default DC font. This is the same as
`WinGrpInfo.NormalFont` – and if WindowSys and GAMEMAIN directly access
this font, so can the endings. Better to keep creation and destruction
in one place, especially with translations in mind.

Closes #20.

Part of P0226, funded by alp-bib and Arandui.