[Maintenance] Remove function calling convention macros
PVOID collides with a nowadays equally useless void* macro from the
Windows headers, so we might as well remove all of them in a single
step. There is little point in micro-optimizing calling conventions and
force-inlining with current compilers: Visual C++ inlines everything
quite aggressively anyway since disk space is cheap nowadays, and
calling convention overrides are ignored in 64-bit builds anyway. Not
to mention that some of the functions with an explicit `__fastcall`
convention didn't take parameters to begin with…

Part of P0246, funded by Ember2528.