[Graphics] [DirectDraw] Change intermediate offscreen bitmaps from DDBs to DIBs
In contrast to DDBs as created by CreateDIBitmap(), DIBs retain the
palette they are initialized with. In the next step, this allows us to
ditch the byte buffer and just retain the HBITMAP, which will then
prepare us for text surfaces. The additional copy from the byte buffer
to the DIB array was there before as well, and is merely made explicit
now.

Part of P0247, funded by Ember2528.