[Config] Implement version-specific views over the config values
I'm amazed. Just a slight bit of advanced modern C++, and we get a
solution that beats a classic polymorphic OOP implementation in all
regards: Flat structures, no class hierarchy, the code is both shorter
and easier to read, and the optimizer can inline *everything*,
resulting in generated code that is not much worse than a manual
straight-line implementation.

Part of P0248, funded by Ember2528, Arandui, and alp-bib.