[3D effects] Line lists: Introduce the concept of world coordinates
POINT is both ambiguous as far as the coordinate space is concerned,
and a Win32 type. Turns out that the WARNING line constants are
interpreted as both a pixel-space and world-space coordinate within the
same structure field. The new initialization is a bit awkward, but does
move the pixel→world-space conversion to compile time. Now we're only
missing the center…

Part of P0249, funded by Ember2528.