[Graphics] [DirectDraw] Use GDI to clear the gradient text bitmap
Should not be graphics backend-specific… and that was in fact the only
instance of a surface being manually cleared, removing yet another
method from the future graphics backend interface.
This also simplifies the color key to a consistent #000000 – no idea
why the palettized mode used a different color.

Part of P0249, funded by Ember2528.