[Graphics] [DirectDraw] Allow color keys to be set explicitly
Very helpful once we start blitting text to images. We want to avoid
the initial upload of a black surface to DirectDraw just to set the
correct color key. We still have to mutate the top-left pixel to get
the correct key value in palettized mode, but at least the usage code
now doesn't have to worry about keeping that pixel in the key color.

Part of P0249, funded by Ember2528.