[Window system] Messages: Additionally keep all text in a single `std::string`
Yup, *additionally*, and redundantly. Eventually, we'd implement
automatic line breaks by splitting this full string into individual
lines at rendering time, but line breaking is complicated and it's way
too early to do this now. Especially because the Stage 4 dialog
actually relies on the wrapping behavior implemented in MWinMsg(),
which retains the first line because it's expected to be the name of
the speaking character.

Part of P0251, funded by Ember2528.