[Graphics] Statically define surface references
Using a separate array of references to avoid adding 10 KB of bloat to
the binary. If we directly initialized references inside `FaceData` and
`EndingGrp`, Visual Studio 2022 would move the entire structure into
the initialized data segment, wasting disk space for a bunch of empty
palettes.
For now, we still track each surface as both a numeric ID and a
SURFACE_DDRAW reference. Migrating to just surface IDs is a bit
premature at this point in time.
This also uncovers the cause behind #19…

Part of P0249, funded by Ember2528.