[Ending] Remove the extra font system
I usually try to keep code that was originally merely commented out,
but this one not only contains a Win32 memory allocation, it also will
be hard to port to non-DirectDraw renderers. I already "preserved" this
one in the form of a video on the ReC98 blog as well.

Completes P0251, funded by Ember2528.