[Score] Replace the key release delay when clicking [END] with a wait state
Not strictly necessary because the main menu blocks against initially
held ⏎ Return and Z keys, but hey, this delay was present in the
original game, so let's keep it. Contrary to the delay when entering
this screen from the main menu, this one does make sense in theory,
since it can show that empty names are turned into "Vivit!". Which it
now does, since we continue to properly render all frames until all
keys are released.

Part of P0255, funded by Ember2528.