[Maintenance] Remove the ZEROMEM() macro
Yeah, nowadays you'd use aggregate initialization for all of these
instances, but that massively bloats the code with these giant
DirectDraw structures. And since it's only used in DirectDraw code, we
can just use the original spelling, and don't have to replace it with
memset().

Part of P0253, funded by Ember2528.