[MIDI] Rewrite the timing algorithm
The original algorithm handled tempo changes by splitting the current
time value in two:
• `playcount1` kept the number of milliseconds *since* the last tempo
  change and increases each time the callback runs, while
• `playcount2` kept the number of MIDI pulses *before* the last tempo
  change, which resets `playcount1` to 0.

This approach necessitated a time→pulse conversion of `playcount1` in
order to calculate when to process the next event, which leads to two
inaccuracies:

• `playcount1` can only be converted to a pulse `C` by combining the
  callback delay with the current tempo. If the MIDI sequence contained
  a tempo change at `B` and before `C`, the new tempo is therefore not
  applied to the MIDI pulses *between* `B` and `C` – it will only take
  effect the next time the callback is called:

  | Tempo: X                    Tempo: Y (won't take effect after `C`)
  |---------------------------------|--------------------------------|
  `A`                              `B`                             `C`
                             MIDI callback delay
  (previous `playcount1`)  +        *            =  (new `playcount1`)
                                 Tempo (X)

• As a consequence, only the latest tempo change between `A` and `C`
  will take any effect. Any additional intermediate tempo changes are
  ignored.

These issues can be avoided by timing MIDI events based on real time
instead of MIDI pulses. Instead of converting the playback time to a
MIDI pulse, we convert every MIDI pulse to milliseconds at perhaps an
excessive level of precision, count down to that event in the callback
by subtracting the delay, and run it once that time is ≤0. By allowing
negative numbers, we naturally allow more than one event per callback,
which is converted using the then current tempo.
However, this poses another challenge with SMF 1 files: What do we do
if one track triggers a tempo change between events on other tracks? In
that case, those tracks preserve the countdown time calculated with the
original tempo, and we've only moved the issue. But this is easily
solved by also tracking the total MIDI pulse of each event, which
allows us to resynchronize all other tracks when such a tempo change
occurs. And since this resynchronization merely needs the difference
between a track's next event and `B`, it even works if other tracks are
multiple events *behind* that tempo change point. I couldn't initially
believe that either.

This shaves roughly 93ms off the full playtime of
「少女神性　～ Pandora's Box」 on my system, which is sadly still 2.6
seconds too slow…?

Part of P0271, funded by Ember2528.