[Build] [Win32] Implement shallow and sparse checkouts of submodules
The combination of which currently still requires setting up the repo
from scratch. In return though, the first checkout we make is already a
sparse one – in contrast to the simple

	git submodule update %module%
	git -C %module% sparse-checkout set --stdin < %module:/=\%.sparse-checkout

where `git submodule update` would still check out the entire repo,
only for most of it to be then deleted by the subsequent command.

Part of P0270, funded by Ember2528.