[MIDI] Retain the current pulse
And show it as the `MIDI TIMER` value in the Music Room, instead of
simply duplicating the `PASSED TIME` and formatting it in milliseconds.
Might be controversial because we no longer reveal the 10ms MIDI update
interval, but I'd say its better to use this UI space to demonstrate
that MIDIs are actually looping.

Part of P0272, funded by Ember2528.