[BGM] Wrap all MIDI functions
These wrappers will handle all the differences between MIDI and
waveform BGM, as well as BGM pack support. Everything that deals with
the playback state of either format has to go here – including MIDI
device changing, which should eventually only restart the currently
playing track if it's a MIDI and not a waveform.

Part of P0272, funded by Ember2528.