[Loader] Asynchronously hash all MIDIs after startup
It may only take 17 milliseconds on my system, but it's something that
I can imagine to be at least noticeable on those old Windows 98 systems
that people will want to run this on…
In contrast to the upcoming loop array, we leave this one in pack order
rather than sorting it by hash value. It will be used for both
file-based and hash-based lookups, and pack order will allow at least
the former to be 𝑂(1).

Part of P0272, funded by Ember2528.