[BGM] [Waveform] Implement linear fades by hand
Yup, miniaudio's fader is linear, and if we wanted to change it to a
logarithmic curve, we'd have to copy-paste a whole lot of highly
redundant C code. Might as well pull that up into the cross-platform
game layer then, where we can implement it in a much more succinct
template.

Also, variants are overrated.

Part of P0273, funded by Ember2528.