[Sound] Add a platform-independent layer
Once we throw BGM into the mix, the BGM and SE subsystems will have to
be initialized and cleaned up independently from each other, using
state management that will be identical between platforms. Kind of
unfortunate that we now have to wrap all the sound effect functions
because I don't want to rename then *again*, but hey, that new
translation unit is also the correct place for Snd_SEStopAll() to be
defined.

Also fixing the mixed-up pause and resume calls in the Win32 event
handler…

Part of P0272, funded by Ember2528.