[Build] Import the C++ standard library as modules
Now with 102% faster compilation times for Debug builds!
Not the order-of-magnitude speedup I've been hoping for, but then again, we're
still including `<windows.h>` all throughout the codebase. Unfortunately, this
also breaks IntelliSense in both VSCode and regular Visual Studio, but the
improved build times are definitely worth it.

Part of P0289, funded by Ember2528.