[Graphics] Backend: Add a generic geometry interface
Alright, the big one. This nicely highlights that the SetColor() and
SetAlpha() methods actually only apply to geometry, not to surfaces.
Also removing the dummy functions for now – we do want to have a null
rendering backend someday, but that's not the way.
Sadly, `AlphaPolygon` can no longer be a raw function pointer once we
have to capture the instances of the upcoming separated channeled and
palettized renderers, which would result in different lambda types.

Part of P0290, funded by Ember2528.