[Graphics] Backend: Add device name retrieval
Unused in the original game, and will remain unused for SDL which
doesn't support either graphics adapter retrieval *or* the ability to
select a specific GPU:

	https://discourse.libsdl.org/t/43659/3

Mainly moving this out now to differentiate it from the upcoming change
that SDL does implement.

Limiting it to 8-bit because the config file didn't reserve more for it
either.

Part of P0290, funded by Ember2528.