[Lens] Capture the pixels under the lens ball into a RAM buffer
pbg's comment indicates that the lens FOV surface was supposed to be in
system memory all along, but the original game didn't make sure to
create it there either. Given that the only purpose of this surface is
to provide temporary storage for pixels while they're getting shifted
around within the locked backbuffer, there's a surefire way of ensuring
a memory surface: Just allocating a regular array and replacing the
DirectDraw blit with a memory copy of the locked pixels. No need to
involve the GPU.
Coincidentally, this also replaces the last piece of DirectDraw in game
code with a solution that will work much better in SDL.

Part of P0289, funded by Ember2528.