[Graphics] [DirectDraw] Don't replicate the palettized color 4 times
It's obvious that this was meant to optimize horizontal drawing by
writing 4 bytes rather than 1 in a single instruction, but then the
code just uses memset() anyway, which only takes a single byte. 🙂
(Not to mention that MSVC's implementation goes much further and can
write 16 bytes at a time using MOVDQA and MOVDQU.)

Part of P0287, funded by Ember2528.