[pbg bug] Fix off-by-one error in Direct3D viewport calculation
`ex` and `ey` are inclusive, so subtracting them from the start
coordinates results in one less pixel than intended. This caused
Direct3D-rendered in-game shapes to be incorrectly clipped to
(511, 479) in the original game, rather than the correct (512, 480)
that DirectDraw-rendered sprites are clipped to.

Part of P0289, funded by Ember2528.