[pbg bug] Fix off-by-one error in 8-bit opaque box drawing
The original game drew these boxes 1 pixel wider than in 16-bit mode.
Given that alpha-blended boxes have the same size in both modes, we've
got enough evidence for a bug. The only opaque boxes drawn in the
entire game are the two red ones behind a newly entered score in the
High Score menu.
Note that changing _2DGrpHline() from inclusive to exclusive would
corrupting the shape of the circle part of long lasers if we don't
simultaneously refactor _2DCircleF(). Annotating what the parameter is
supposed to be is a good idea in any case.

Part of P0289, funded by Ember2528.