[Graphics] [Direct3D] Defer viewport updates until locking
Sure, IDirect3DDevice2::SetViewport() takes only a low 4-digit number
of nanoseconds, but it's measurable. This optimization doesn't help as
much as you'd expect because CEffectDraw() locks the Direct3D surface
during each in-game frame anyway, but it will counteract the fact that
we're going to have to call Grp_SetClip() once per frame for the HUD
text.
As a side effect of the necessary caching, we also get to limit the Z
buffer clears to the viewport region, in case it matters.

Part of P0291, funded by Ember2528.