[Maintenance] Start a cross-cutting header file for game-specific constants
Defining the two sound-related ones in only a single translation unit
wasn't too bad. However, it's going to get much worse once we start
doing the same for the native resolution, where we have the problem of
`Grp_AddrY[480]` referring to a game-specific constant at the platform
level.
There are a number of ways to handle this. But given that every layer
would like to use this file *and* that the platform layer should
*theoretically* be reusable, including these constants from an
unspecified directory (and burdening the build system with it) seems
like the least bad option.

Part of P0289, funded by Ember2528.