[Text] [GDI] Move the text surface reference into the graphics backend
After all, it's this backend that decides how the additional GDI
surface is handled. DirectDraw can simply use the GDI part of the
DirectDraw surface, whereas SDL needs to declare a separate
`SURFACE_GDI`. Trying to handle this difference in game code quickly
runs into a static initialization order fiasco as we can't pass
`DxSurf[SURFACE_ID::TEXT]` into the `TextObj` constructor. Sure, those
three `#ifdef`'d GDI-related functions are rather unsightly, but it
sure is better than solving this problem with another abstraction.

This also reveals how SurfaceEdit() was the wrong abstraction – since
text rendering is the only foreseeable case where we want to update a
backend surface with the contents of a GDI surface, we might as well
spell out these GDI operations as such. Image patching for translations
will probably be done on the BMP buffer; if we ever need a
color→palette matching algorithm, it's better to write a custom one at
game scope instead of burdening the graphics backend.

Part of P0291, funded by Ember2528.