[Graphics] Game-level geometry generation: Add gradient rectangles
Since these use two colors, we will have to generate the whole set of
colors for every vertex *somewhere*. Since this layer doesn't know the
current geometry color (and really shouldn't), it makes more sense to
express that one as a parameter and keep the white color hardcoded in
this new geometry layer. Inconsistent with the rest of the code, for
sure, but a great first step into a less stateful codebase. The laser
bomb code already benefits quite a bit from this.

Part of P0290, funded by Ember2528.