[Window system] Menus: Don't dedent titles
Maybe it would be cleaner to instead indent every other string, but
that would leave us with some quite ugly `sprintf()` calls. Fixes the
title of the `Joy Pad` menu (again, why did I not bother in 4ecce21?)
in addition to the submenus we're about to add.

Part of P0292, funded by Ember2528.