[MIDI] Add an SC-88Pro compatibility option for patching invalid Reverb Macros
Romantique Tp confirmed that real hardware clamps invalid messages:

	https://twitter.com/Romantique_Tp/status/1766895996645056902

We can cover the AST MIDIs within the BGM packs by simply modifying
them, but we can't do the same for the OST MIDIs due to copyright.
Therefore, the best way of offering SC-88Pro compatibility is this
optional hotpatch directly in the MIDI subsystem. It might seem as if
this particular fix fits better at the game-specific level, right after
loading the MIDI file, but:

• That would require publicly accessible MIDI event iteration, because
  we definitely ain't going to loop over the raw bytes ourselves
• The `MID_EVENT` spans could no longer be `const`
• And even then, it would only work because this specific patch doesn't
  resize the message. It's very questionable to implement any kind of
  MIDI hot-patching with this serious limitation…
• …especially if features like drum note lengthening (#9) and
  reflavoring (#10) couldn't work within this limitation, and thus
  require the same kind of hardcoded patching.