[Enum flags] Stay within the class type after `operator &`
Less surprising and more friendly for using this operator outside of a
boolean context… at the cost of requiring another trick *for* said
contexts. Overloading `!!` as a replacement for the impossibility of
overloading `operator bool` might be controversial, but it might also
be just fine given that we don't use this idiom for anything else.

Part of P0292, funded by Ember2528.