[Replays] Only backup the three overridden config fields
The upcoming hotkey feature allows config fields to be changed in-game,
and we obviously don't want to reset them after a replay finished
playing.
And given that we fixed the `GameLevel`-related bugs earlier, we don't
even have to backup that one.

Part of P0295, funded by Ember2528 and [Anonymous].