[Enemy] Remove casts from `ID2Value()`
The original `DWORD` obviously meant "largest type that can fit any of
the values". Compared to the alternative of using an `union` with one
field for every type, this is definitely the most concise solution for
the problem, not least thanks to the guaranteed zero extension of
smaller values. Any explicit casting here is pointless to begin with.

The only sensible measure is to prevent future modders from returning
64-bit values by opting into the associated error. We sure ain't going
to enable this warning globally, I've had enough of this by now…

Part of P0287, funded by Ember2528.