[Config] Store all fundamental types as big-endian
We've been using big-endian for string lengths as another test case for
`game/endian.h`, so it makes sense to also use it for the upcoming
>8-bit values. However, we only actually want it for storage. The
in-memory type should still be the native one, not least because MSVC's
optimizer still struggles to detect the byte swap.

Part of P0292, funded by Ember2528.