[Build] Move all Win32-exclusive Tupfile code to a separate file
The Linux build will change so much that it isn't worth it to somehow
parametrize it within the current script. The only thing that we can
meaningfully share between Windows and Linux are these include and
object paths, the `PBG_DEBUG` macro, and the globs for the game logic
code.
(Thankfully, Visual Studio supports the dashed option syntax as well.)

Part of P0300, funded by Ember2528.