[Path] [SDL] Allow the root of `$XDG_DATA_HOME` to be used as the data path
Very useful for Flatpak, which sets $XDG_DATA_HOME to the already
app-specific and non-overridable `~/.var/app/<app-id>/data`. It would
seem that you could just add

	--filesystem=xdg-data/<org>/<app>
	--env=XDG_DATA_HOME=~/.local/share/<org>/<app>

to the `finish-args`, but Flatpak deliberately prevents this from
working:

	https://github.com/flatpak/flatpak/issues/2413

(Not to mention that the resulting manifest would then fail the
`finish-args-unnecessary-xdg-data-subdir-mode-access` lint and couldn't
be published on Flathub without arguing for an exception. Good luck
with that.)

Part of P0303, funded by Ember2528.