[Build] Move the submodule error to a separate template file
We'd like to show it on Linux as well.
Using colons to separate `sed` parameters is a great idea here: Module
paths create a problem with slashes, but colons are illegal in
filenames on Windows, so we won't ever have them in valid paths inside
this repository.

Part of P0302, funded by Ember2528.