[Build] Enable full dependency tracking for Linux
MSVC compiler and C++ runtime updates typically come with a change to
at least one environment variable within the Visual Studio command
prompts, which allows Tup to catch them and fully rebuild everything.
But this does not apply to Linux compilers, where `full_deps` is the
only way to catch these changes. This is especially crucial for modules
on Clang because BMIs are tied to a specific compiler version and will
throw weird errors after an update.
And yes, that will be a Tup PR somewhere down the line.

Part of P0302, funded by Ember2528.