[Graphics] [SDL] Keep OpenGL version strings regardless of SDL's support macros
SDL 3 turned these into build time-exclusive macros and removed them
from the public headers:

	https://github.com/libsdl-org/SDL/issues/6643
	https://github.com/libsdl-org/SDL/commit/63f307fe1fa662dbf2ffb2d2194d2d8b3191257d

And they're right – the supported backends are a property of the SDL
DLL/SO, which is supposed to be swappable without recompiling the game.
This micro-optimization might seem like the clean thing to do if you
squint at it from a certain angle. But it never actually hurts to
define those strings, and these `#ifdef`s would have only *actually*
removed those few bytes in a hypothetical Windows 98 build without
OpenGL.

A few days ago, Arch Linux was maybe a bit too reckless and replaced
regular SDL 2 with sdl2-compat on top of SDL 3. sdl2-compat defines a
few of these macros for compatibility reasons with similarly misbehaved
programs:

	https://github.com/libsdl-org/sdl2-compat/blob/main/include/SDL2/SDL_config.h#L60-L87

But it can't know all the `SDL_VIDEO_RENDER_*` ones we are using here,
which therefore caused the build to fail due to an empty `Versions`
array.