[Platform] Add SDL 3 support
For now, we still must support SDL 2 for the vintage Windows binary
until we've backported SDL 3 to Windows 98. Keeping support for a few
more years is also a good idea for for more conservative Linux
distributions, though.
Going for the hybrid approach of targeting both versions within the
same general code, at the cost of 59 little branches.

Closes #68.

Part of P0307, funded by Ember2528.