[Maintenance] Consistently format preprocessor branches
Always placing the directives in the first column and not adding
further indentation to the branch body, which matches the convention
that a part of the code already used. This makes indentation exclusive
to run-time branches, which will work even better once we have a
`compile_commands.json` for every configuration.
(Also, clang-format prefers this as well.)

Part of P0307, funded by Ember2528.