[Graphics] Screenshots: Create the file immediately before writing to it
Sure, this might cause the backend to needlessly capture the screen if
the file write fails due to an I/O error. But once we add WebP support,
it'll be much more important for the frontend to dynamically pick the
file format. And besides, WebP encoding will be much slower than the
pixel capture.

Part of P0308, funded by Ember2528.