[Sound] [miniaudio] Migrate to miniaudio.c
miniaudio will move the implementation out of the header in version
0.12. Using the split version is not much more complicated if you stay
focused on the code and not the build system. Expressing this in Tup
land would be 24 lines of dense branching code if you don't want
miniaudio's flags to pollute the other platform build command lines.

Part of P0307, funded by Ember2528.