[Music Room] Render all text using the immediate interface
Retained text rendering fundamentally does not work when hotkeys can
change graphics params and cause the text surface to be wiped at any
time. This already caused a bug in the Music Room on GDI:

1) Enter the Music Room in windowed mode
2) Change the window size using the F10 and F11 hotkeys. The text
   surface now gets wiped on the next (pre)rendering call.
3) Move to another track
4) All prerendered text (i.e., the version banner and the MIDI device)
   is now gone as the title/comment prerendering call will now wipe the
   surface and the old text won't get prerendered again.

Sure, we might not need to wipe the text surface after *all* graphics
param changes as we do now, but we *will* have to wipe it if we ever
implement #72. So let's just remove the retained interface entirely.

Part of P0308, funded by Ember2528.