[Ending] Render using the immediate interface
No bugs there as there only is a single rectangle that must be
completely re-rendered on every text change. (There's one box that adds
a new line to a box without blanking it, which changes the horizontal
position of the line before).
And now that the last retained instance is rewritten, we can…

Part of P0308, funded by Ember2528.