[Build] Generate a macro for the current Git commit and tag
Basically a POSIX shell port of the same code for the ReC98 Anniversary
Editions:

	https://github.com/nmlgc/ReC98/commit/87dc80711c52a6363ac04c5ffd7af5ea13dd0207

With the exception that we don't use the full `git describe` string and
instead just display the commit limited to around 4 characters, as we
won't have all that much screen space right of the help text window.

Part of P0308, funded by Ember2528.