[Window] [SDL] Make renderer name lookup more resilient
SDL's renderer choice is well defined in the absence of a
`SDL_RENDER_DRIVER` hint, pretending it isn't only causes harm in case
the hint doesn't exist or is invalid. (Specifically, crashing the API
submenu because it subsequently tries to do strcmp() on a `nullptr`.)

Part of P0308, funded by Ember2528.