[Graphics] [SDL] Move the default API from a hint to a constant
SDL won't set a hint at default priority if the corresponding
environment variable is set, and we'd like a consistent default for
([params.api] < 0) regardless.
And yes, a pointer has much better ergonomics than a string_view if we
want it to be nullable *and* have logic that requires constant checks
anyway.

Part of P0308, funded by Ember2528.