[Window system] Merge states into flags
We might want to have items that are both disabled and highlighted. And
now that EnumFlagSet() makes this much more comfortable, it becomes
obvious how much of a bad idea this whole split was to begin with.
After all, flag changes should also trigger a re-render.

Part of P0309, funded by Ember2528.