[Graphics] [SDL] Don't indicate the active renderer by copying pointers
If we reinitialized the primary renderer while software-rendering, the
`Renderer` pointer would previously no longer indicate that we were
software-rendering. Sure, we could have reinitialized `Renderer`
depending on the value of the `SoftwareTexture`, but that would have
been yet another branch that relied on global state.

Part of P0308, funded by Ember2528.