[Bug] Replays: Write actual difficulty and lives to the header
Ever since 47f365a, we've been saving Extra Stage replays with the
configured difficulty rather than Hard, which caused newly recorded
replays to desync if the difficulty was set to any other value. (Which
it probably was.) Same for the configured player stock rather than the
fixed 2.
Thanks to >>49320040 on /jp/ for pointing out that desyncs exist, that
is not supposed to happen. Please tell me this sort of thing! I'm not
ZUN, desyncs are critical bugs that definitely fall under my free
bugfix policy, and will receive my immediate attention.