⚠️ [Gameplay regression] Fix wrong cast in Marisa's bit direction calculation
Yup, it's the dreaded case of a small oversight causing an accidental
gameplay fork. While it is common for Extra Stage playthroughs to run
into a case where this wrong cast calculates a different `dir` value,
this seems to only cause actual desyncs if the ECL script then branches
to a different pattern depending on the value it calculated from the
`d` variable of the associated enemy. I have 6 Extra Stage replays that
beat Marisa, and none of them desync in response to this bug.

Many, many thanks to Ripper Roo for reporting this bug and providing
the apparently rare kind of replay that reproduces the issue. Luckily,
that replay was recorded on P0275 and thus expected the original game's
behavior that this commit restores. Let's hope that there are no
replays out there that were recorded on ≥P0295 and rely on the wrong
behavior, because we most likely can't do anything about *those*
desyncs…