[Sound] Load .WAV files in the frontend, using SDL
Removes both miniaudio's integrated dr_wav *and* the `fopen()`
implementation it pulled from the standard library, saving 54 KiB in
the modern build and 45.5 KiB in the vintage build! In exchange for 25%
of performance…
Also applied the change to the DirectSound backend in case we ever
decide to get rid of miniaudio (or even just its resampler), where it
hilariously removes much more lines of code than it adds. Using that
backend right now requires some hacking to re-expose the Win32 handle;
I'll properly solve that when we actually make use of it again.

Closes #24.

Part of P0326, funded by Ember2528.