[Sound] [miniaudio] Disable SSE2 and AVX2 for the vintage build
These now seem to cause an unskippable debug break on `__cpuid()`? Not
optimal for people who want to use this build on XP, I get it. But we
don't want to have *any* ≥SSE instructions in the vintage binary, and
the easiest way to prove their absence is to purge all of them.

Part of P0324, funded by Ember2528 and Root.