[Maintenance] Remove pbg's Windows-native window, path, and DirectInput code
Now that we're all in on SDL and being cross-platform, we're no longer
going to maintain this code, and it will only bitrot going forward. The
DirectSound backend remains valuable for resampling verification
reasons if we ever address #50, but the rest doesn't contain anything
of value anymore.

Part of P0324, funded by Ember2528 and Root.